fixup! Issue #1881 - Pass extras from launch Intent through to Gecko in dev builds.
